app/requirements.txt
flask==3.0.3
flask-sqlalchemy==3.1.1
loguru==0.7.2
python-dotenv==1.0.1